移植 26.2 Fabric

+++ src/main/java/org/mesdag/particlestorm/data/molang/compiler/function/generic/AbsFunction.java
@@ -4,11 +4,14 @@ import org.mesdag.particlestorm.api.MolangInstance;
 import org.mesdag.particlestorm.data.molang.compiler.MathValue;
 import org.mesdag.particlestorm.data.molang.compiler.function.MathFunction;
 
-/// [MathFunction] value supplier
-///
-/// **Contract:**
-///
-/// Returns the absolute (non-negative) equivalent of the input value
+/**
+ * {@link MathFunction} value supplier
+ *
+ * <p>
+ * <b>Contract:</b>
+ * <br>
+ * Returns the absolute (non-negative) equivalent of the input value
+ */
 public final class AbsFunction extends MathFunction {
     private final MathValue value;
 
@@ -24,7 +27,7 @@ public final class AbsFunction extends MathFunction {
     }
 
     @Override
-    public float compute(MolangInstance instance) {
+    public double compute(MolangInstance instance) {
         return Math.abs(this.value.get(instance));
     }
